修正代码

+++ guns/guns-admin/src/main/java/com/stylefeng/guns/modular/bank/controller/BankController.java
@@ -212,7 +212,15 @@ public class BankController extends BaseController {
      */
     @RequestMapping(value = "/delete")
     @ResponseBody
-    public Object delete(@RequestParam Integer bankId) {
+    public Object delete(@RequestParam Integer bankId) throws Exception {
+        Bank bank = bankService.selectById(bankId);
+        EntityWrapper<BankDeposit> wrapper = new EntityWrapper<BankDeposit>();
+        wrapper.like("bankNo", bank.getBankNo());
+        List<BankDeposit> bankDeposits = bankDepositService.selectList(wrapper);
+        if (CollectionUtils.isNotEmpty(bankDeposits)) {
+            throw new Exception("银行卡：" + bank.getBankNo() + "，存在流水数据数据，不允许删除");
+        }
+
         bankService.deleteById(bankId);
         return SUCCESS_TIP;
     }
@@ -226,6 +234,12 @@ public class BankController extends BaseController {
         if (StringUtils.isEmpty(bank.getBankNo())) {
             throw new Exception("请输入银行卡号");
         }
+        //查询一下是否存在流水，如果存在并且卡号发生变化则更新
+        Bank temp = bankService.selectById(bank.getId());
+        if (!bank.getBankNo().equals(temp.getBankNo())) {
+            bankDepositService.updateName(bank.getBankNo(), temp.getBankNo());
+        }
+
         if (bank.getBalance() == null) {
             bank.setBalance(new BigDecimal(0));
         }
